Refactor: реализация CustomLinkedList (методы linkLast, getTasks, removeNode)

+++ src/manager/HistoryManager.java
@@ -10,5 +10,5 @@ public interface HistoryManager {
 
     void add(Task task);
 
-    void remove(int id);
+    void remove(long id);
 }
