+++ src/task/Subtask.java
@@ -2,14 +2,14 @@ package task;
 
 public class Subtask extends Task {
 
-    private int epicId;
+    private long epicId;
 
-    public Subtask(String nameTask, String description, int id, StatusTask statusTask, int epicId) {
+    public Subtask(String nameTask, String description, long id, StatusTask statusTask, long epicId) {
         super(nameTask, description, id, statusTask);
         this.epicId = epicId;
     }
 
-    public int getEpicId() {
+    public long getEpicId() {
         return epicId;
     }
